+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Added
 
 - Now create README.md in meta directory to explain how to use it.
+- [#2](https://github.com/creekorful/srcode/issues/2) Implement srcode ls.
 
 ## [0.1.0] - 2021-01-15